+++ commercial/app/views/contentapi/item.scala.html
@@ -5,7 +5,7 @@
 @import views.html.fragments.commercial.cards.itemLargeCard
 @import views.html.fragments.commercial.containerWrapper
 @import views.html.fragments.inlineSvg
-@import views.support.commercial.TrackingCodeBuilder.mkCapiCardTrackingCode
+@import views.support.Commercial.TrackingCodeBuilder.mkCapiCardTrackingCode
 @import views.support.Commercial.glabsLink
 
 @(card: CardContent,
