+++ common/app/views/fragments/commercial/cards/itemLargeCard.scala.html
@@ -3,7 +3,7 @@
 @import views.html.fragments.commercial.cardLogo
 @import views.html.fragments.inlineSvg
 @import views.html.fragments.items.elements.facia_cards.image
-@import views.support.commercial.CssClassBuilder
+@import views.support.Commercial.CssClassBuilder
 
 @(item: common.commercial.CardContent,
   omnitureId: String,
