+++ common/app/views/fragments/commercial/cards/itemSmallCard.scala.html
@@ -6,7 +6,7 @@
 
 @import views.html.fragments.commercial.cardLogo
 @import views.html.fragments.inlineSvg
-@import views.support.commercial.CssClassBuilder
+@import views.support.Commercial.CssClassBuilder
 
 <a href="@item.targetUrl"
    class="@CssClassBuilder.linkFromSmallCard(item, optAdvertClassNames, optClassNames, useCardBranding)"
