+++ common/app/views/fragments/commercial/containers/paidContainer.scala.html
@@ -7,7 +7,7 @@
 @import views.html.fragments.commercial.containers._
 @import views.html.fragments.commercial.containerLogo
 @import views.html.fragments.inlineSvg
-@import views.support.commercial.TrackingCodeBuilder.mkInteractionTrackingCode
+@import views.support.Commercial.TrackingCodeBuilder.mkInteractionTrackingCode
 @import views.support.Commercial.glabsLink
 
 <div data-id="@containerModel.id" class="fc-container">
